requirements.app.txt
google-genai>=1.51.0
google-cloud-texttospeech>=2.28.0
Pillow>=10.4.0
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
google-cloud-storage>=3.1.0
python-dotenv>=1.0
